requirements.txt
Flask==3.0.0
pdf2image==1.16.3
pytesseract==0.3.10
Pillow==10.1.0
img2pdf==0.6.3

